C ve D, 2 seçenekli rol için düzenlemeler

+++ src/Controllers/PageController.php
@@ -146,16 +146,6 @@ class PageController extends Controller
             $page_permissions[] = [
                 'role_id' => $r,
                 'page_id' => $page->id,
-                'permission' => 'R',
-            ];
-            $page_permissions[] = [
-                'role_id' => $r,
-                'page_id' => $page->id,
-                'permission' => 'U',
-            ];
-            $page_permissions[] = [
-                'role_id' => $r,
-                'page_id' => $page->id,
                 'permission' => 'D',
             ];
         }
@@ -190,7 +180,9 @@ class PageController extends Controller
      */
     public function edit(Page $page)
     {
-
+        if(!Auth::user()->hasPagePermission($page->id,'C')){
+            return abort(403);
+        }
         $detail_extra_include = false;
         $extra_include = false;
         if(View::exists('vendor.prime.extras.page'.$page->id.'detail-extra.main')){
@@ -276,6 +268,9 @@ class PageController extends Controller
 
     public function update(Request $request, Page $page)
     {
+        if(!Auth::user()->hasPagePermission($page->id,'C')){
+            return abort(403);
+        }
         $order_check = null;
         if ($request->has('order') && $page->page_id) {
             $this->setOrders($request->post('order'),$page);
@@ -363,6 +358,9 @@ class PageController extends Controller
      */
     public function destroy(Page $page)
     {
+        if(!Auth::user()->hasPagePermission($page->id,'D')){
+            return abort(403);
+        }
         $page->delete();
         $this->createLog($page,Auth::user()->id,"D");
         return redirect()->route('pages.index');
